apps/worker/requirements.txt
pydantic
python-dotenv
psycopg[binary]
SQLAlchemy
beautifulsoup4
pypdf
tiktoken
